aded size

+++ output_integer.c
@@ -11,8 +11,8 @@ int output_integer(const char *s, va_list arg)
 {
 	int a[16];
 	long int j = 1, m = 1000000000000000, n;
-
-	long int counter = 0, sum = 0;
+	
+	int counter = 0, sum = 0;
 
 	n = va_arg(arg,int);
 	while (*s != 'i')
